Added member ship type  drop down to the form.
This included addition of a view model. NewCustomerViewModel. It contains a customer object and a MemberShipType IEnumenrable. This is done so that it can be passed to the view as one single package in the New action from within the Customer controller. Also, have had to add the MemberShipTypesID to db set within the IdentityModels class.